requirements.txt
Django==4.2.16
psycopg2-binary==2.9.9
Pillow==10.4.0
crispy-bootstrap5==2024.2
django-crispy-forms==2.3
whitenoise==6.7.0
python-decouple==3.8
